8578 - We can't add new things to an older update
Each update function is expected to only be run once or not at all, depending
on when the site was installed.

Drupal uses the number at the end of the function as a version number for each
module. On install, it finds the highest number defined and uses that for the
version number. It will refuse to run any update with a number less than or
equal to the version number.

This update has already been run, and the version number for `os_dataverse` is
set to 7002. We need to make a new update function (7003) to do any additional
work.