8562 - Redo "Read More" link changes per @rbran100
Revert out all of the previous readmore changes thus far in the new branch
`8562-readmore-redo`, ... the previous solution will cause more bugs in the
future.

The readmore issue (not showing for profiles) was specific to themes extending
`hwpi_basetheme`.

We can see what it is doing in `node_view_alter` here:
  https://github.com/openscholar/openscholar/blob/SCHOLAR-3.x/openscholar/themes/hwpi_basetheme/template.php#L194

We noticed that and removed the `unset` on the links which was exactly
right:
  https://github.com/openscholar/openscholar/pull/8619/files#diff-4773c00a5c8f1dc142ee969856ca0018R257

However, above that (line 207), the `['body']` is completely blown away and moved
inside of `pic_bio`. This breaks all of our processing of the `body` field for
readmore links. We have code that only displays the readmore when the body is
actually long enough but that is not working since `hwpi_basetheme` blows `body`
away.

So, how to solve this...

  1. Getting rid of the `unset` was perfect, redo that

  2. We need to have a way of putting the body inside of a `pic_bio` wrapper,
    without removing it. What we need to do, is create that `bio_pic` wrapper and
    move the needed elements inside of it, later in rendering, during
    `hwpi_basetheme_process_node` instead of `hwpi_basetheme_preprocess_node`. Both
    functions already exist, it should just be a matter of moving the wrapper, and
    making sure things get rendered correctly. We can compare the markup before
    and after and it should be the same.... except now the read-more links will
    show.