9018 - ensure trailing slash on URLs #9018
Adding the trailing slash makes relative links relative to the vsite level,
instead of the SERVER_NAME level.

Note: a link that was previously "dependant" on a URL *without* a trailing
slash will need to be pre-pended with a "../" to point to it's old target.
Alternatively, users needing a link relative to the vsite directory level,
could be required to prepend links with the vsite name.