9018 - revert previous commit (don't touch $_ globals) #9018
Changing the REQUEST_URI like this is a non-starter. You shouldn't be changing
the values of $_ variables.  We should be parsing relative links in php to make
them absolute by the time they hit the client.  Changing the REQUEST_URI, or
any $_SERVER variable, is not the way to go.